+++ connector/src/main/java/it/eng/connector/configuration/InitialDataLoader.java
@@ -131,7 +131,8 @@ public class InitialDataLoader {
                         .build()
                         .toString();
 
-                s3ClientService.uploadFile(FileUtils.openInputStream(file.getFile()), bucketName, fileKey, MediaType.APPLICATION_JSON_VALUE, contentDisposition)
+                s3ClientService.uploadFile(FileUtils.openInputStream(file.getFile()), bucketName, fileKey,
+                                MediaType.APPLICATION_JSON_VALUE, contentDisposition)
                         .get();
             }
         } catch (Exception e) {