+++ tools/src/main/java/it/eng/tools/s3/configuration/S3Configuration.java
@@ -1,5 +1,6 @@
 package it.eng.tools.s3.configuration;
 
+import io.minio.admin.MinioAdminClient;
 import it.eng.tools.s3.properties.S3Properties;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Bean;
@@ -19,13 +20,13 @@ import java.util.concurrent.Executors;
 @Configuration
 @Slf4j
 public class S3Configuration {
-    
+
     private final S3Properties s3Properties;
-    
+
     public S3Configuration(S3Properties s3Properties) {
         this.s3Properties = s3Properties;
     }
-    
+
     @Bean
     public S3Client s3Client() {
         return S3Client.builder()
@@ -36,6 +37,14 @@ public class S3Configuration {
                 .serviceConfiguration(software.amazon.awssdk.services.s3.S3Configuration.builder()
                         .pathStyleAccessEnabled(true)
                         .build())
+                .build();
+    }
+
+    @Bean
+    public MinioAdminClient minioAdminClient() {
+        return MinioAdminClient.builder()
+                .endpoint(s3Properties.getEndpoint())
+                .credentials(s3Properties.getAccessKey(), s3Properties.getSecretKey())
                 .build();
     }