+++ tools/src/main/java/it/eng/tools/s3/service/IamUserManagementService.java
@@ -1,0 +1,8 @@
+package it.eng.tools.s3.service;
+
+public interface IamUserManagementService {
+
+    void createUser(BucketCredentials bucketCredentials);
+
+    void attachPolicyToUser(BucketCredentials bucketCredentials);
+}