+++ CHANGELOG.md
@@ -8,10 +8,13 @@ All notable changes to this project will be documented in this file.
 
 - When creating S3 bucket, a policy for restricting access will be created
 - S3 client providers, for creating clients with different configurations
+- IAM user management for S3 bucket access
+- minio-admin dependency for managing IAM users for MiniIO
 
 ### Changed
 
 - minio version minio/minio:RELEASE.2025-04-22T22-12-26Z - has extended UI interface
+- updated tests
 
 ## [0.4.3-SNAPSHOT] - 20.06.2025.